Create an abstract mixin config plugin
