--- src/main/java/de/kb1000/notelemetry/CommonUtil.java
@@ -1,23 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * This Source Code Form is "Incompatible With Secondary Licenses", as
- * defined by the Mozilla Public License, v. 2.0.
- */
-package de.kb1000.notelemetry;
-
-class CommonUtil {
-    static boolean isForgeNeo() {
-        return classExists("net.neoforged.fml.common.Mod");
-    }
-
-    static boolean classExists(String name) {
-        try {
-            return NoTelemetryFabricMixinConfigPlugin.class.getClassLoader().loadClass(name) != null;
-        } catch (Exception | LinkageError e) {
-            return false;
-        }
-    }
-}
